check every robot instead of using book request
To be decided if it is better, by checking just the books there are a
lot less of requests, but the robots will not be displayed as "Active".
In contrast with this setup if there are too many robots they will not
all be checked every 2 minutes so also will not be displayed as
"Active". Ideally it would be better that when creating a robot it would
be possible to set an option always_online and then using book requests.